requirements.txt
torch>=2.1.0
transformers>=4.45.0
datasets>=2.14.0
accelerate>=0.25.0
huggingface-hub>=0.20.0
safetensors>=0.4.0
click>=8.0.0
python-dotenv>=1.0.0
numpy>=1.26.0
matplotlib>=3.8.0
